Chore/upgrade rails7

See merge request am/levumi!4

+++ app/controllers/tests_controller.rb
@@ -1,47 +1,56 @@
 class TestsController < ApplicationController
-  before_action :set_test, except: [:index, :create]
-  before_action :is_allowed, only: [:create, :edit, :update, :destroy]
+  before_action :set_test, except: %i[index create get_tests_data]
+  before_action :is_allowed, only: %i[create edit update destroy]
 
   #GET /tests
   def index
-    if params[:admin] && @login.has_capability?('test')
-      render 'index_admin'
+    #TODO check where which params are used.
+    if params[:show_export] && @login.has_capability?('export')
+      render 'index'
     else
-      if params[:export] && @login.has_capability?('export')
-        render 'index_export'
-      else
-        render 'index'
-      end
+      render 'list'
     end
+    #if params[:admin] && @login.has_capability?('test')
+    #  render 'index_admin'
+    #else
+    #  if params[:export] && @login.has_capability?('export')
+    #    render 'index_export'
+    #  else
+    #    render 'index'
+    #  end
+    #end
   end
 
   #GET /tests/:id/edit
-  def edit
-  end
+  def edit; end
 
   #GET /test/:id
   def show
     #TODO: Sicherheitscheck ob erlaubt?
-    send_data @test.as_csv, filename: @test.shorthand + '_' + DateTime.now.strftime("%Y_%m_%d") + '.csv' , type: :csv
+    send_data @test.as_csv,
+              filename: @test.shorthand + '_' + DateTime.now.strftime('%Y_%m_%d') + '.csv',
+              type: :csv
   end
 
   #POST /tests
   def create
     if params.has_key?(:test) && !params[:test][:file].nil?
-      params[:test][:file].each do |f|
-        @import_failure = Test.import(f.tempfile, !params.has_key?(:update_test), params.has_key?(:update_material)).nil?
+      f = params[:test][:file]
+      @import_failure =
+        !f.present? ||
+          Test.import(f.tempfile, !params.has_key?(:update_test), params.has_key?(:update_material))
+            .nil?
+      if @import_failure
+        head :unprocessable_entity
+      else
+        head :ok
       end
-      render 'index_admin'
     end
   end
 
   #PUT /tests/:id
   def update
-    if !@test.update_attributes(test_attributes)
-      render nothing: true
-    else
-      render 'update'
-    end
+    render nothing: true if !@test.update(test_attributes)
   end
 
   #DEL /tests/:id
@@ -49,29 +58,28 @@ class TestsController < ApplicationController
     id = @test.test_family.id
     @test.destroy
     family = TestFamily.find(id)
-    if family.tests.empty?
-      family.destroy
-    end
-    render 'index_admin'
+    family.destroy if family.tests.empty?
+    head :ok
+  end
+
+  def get_tests_data
+    @data = Area.includes(competences: [{ test_families: [:tests] }]).all
   end
 
   private
 
   #Erlaubte Attribute definieren
   def test_attributes
-    params.require(:test).permit(description: [:full, :short])
+    params.require(:test).permit(description: %i[full short])
   end
 
   #Prüfen ob Nutzer die Berechtigung für Testaktualisierungen hat
   def is_allowed
-    unless @login.has_capability?('test')
-      redirect_to '/'
-    end
+    redirect_to '/' unless @login.has_capability?('test')
   end
 
   #Test laden
   def set_test
     @test = Test.find(params[:id])
   end
-
 end
